+++ src/Counter.jsx
@@ -1,22 +1,12 @@
-import NumberButton from './NumberButton';
 import ClickMeButton from './ClickMeButton';
-
-const NUMBERS = [1, 2, 3, 4, 5];
+import NumberButtons from './NumberButtons';
 
 export default function Counter({ count, clickHandler }) {
   return (
     <>
       <h1>Counter</h1>
       <ClickMeButton count={count} clickHandler={() => clickHandler(1)} />
-      <div>
-        {NUMBERS.map((number) => (
-          <NumberButton
-            key={number}
-            number={number}
-            clickHandler={() => clickHandler(number)}
-          />
-        ))}
-      </div>
+      <NumberButtons clickHandler={clickHandler} />
     </>
   );
 }
